backend/requirements.txt
# Core
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# Pose estimation
mediapipe>=0.10.0
numpy>=1.24.0,<2.0.0
opencv-python>=4.8.0

# Utilities
pydantic>=2.0.0
python-dotenv>=1.0.0

# Gemini AI
google-genai>=1.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
httpx>=0.24.0
